Edits
LimelightLogger -> VisionLogger
Now in its own file and added some logging into periodic

Pilot.java
removed a line to not mess with Pilot.java in main.
