+++ src/main/java/frc/robot/pilot/Pilot.java
@@ -29,7 +29,7 @@ public class Pilot extends Gamepad {
     public final Trigger actionReady = rightBumper.and(teleop);
 
     // vision Drive
-    public final Trigger visionAim = Y.and(noFn, teleop);
+    // public final Trigger visionAim = Y.and(noFn, teleop);
     // Drive Triggers
     public final Trigger upReorient = upDpad.and(fn, teleop);
     public final Trigger leftReorient = leftDpad.and(fn, teleop);
